refactor: DB pool 호출위치 변경
