+++ dao/admin.js
@@ -1,5 +1,4 @@
-const pool = require('./db');
-
+const pool = require('../configs/mysql');
 const { databaseError } = require('../utils/errors/database');
 
 const getUserList = async () => {
